+++ Lab2_Alterno/laboratorio2.simu
