--- Lab2_Alterno/main.hex
@@ -1,131 +1,0 @@
-:1000000014C0FBC02CC12CC02BC02AC0D9C028C032
-:1000100027C026C025C055C123C022C021C020C092
-:100020001FC01EC01DC090C11BC011241FBECFE544
-:10003000D1E0DEBFCDBF10E0A4E6B0E001C01D926C
-:10004000A03BB107E1F710E0A0E6B0E0EEE0F8E099
-:1000500002C005900D92A436B107D9F716D2D5C3C8
-:10006000CFCF22E029BF21E030E0309387002093FA
-:100070008600109283001092820012BE1816190694
-:1000800024F488B3809588BBFCCF18BA19BE1092AF
-:1000900087001092860008951F9380918A009091A6
-:1000A0008B00019769F41FE018BB8AEF90E0D9DF5D
-:1000B00018BA8AEF90E0D5DF18BB8AEF90E00EC047
-:1000C0001FE018BB80ED97E0CCDF18BA18BB80EDBD
-:1000D00097E0C7DF18BA18BB80ED97E0C2DF18BA07
-:1000E0001F910895CF92DF92EF92FF920F931F938B
-:1000F000CF93DF9306E610E0C0E0D0E024E0C22E0C
-:1001000092E0D92EEE24E39488E0F82E22C0F80184
-:10011000808191818230910571F0833091051CF4CA
-:100120000197A1F406C08330910539F0049771F46A
-:1001300006C0F8BA05C0C8BA03C0D8BA01C0E8BA48
-:100140008091AE009091AF008CDF18BA21960E5FBF
-:100150001F4F8091840090918500C817D907BCF289
-:100160008091AE009091AF00885C90409093AF007A
-:100170008093AE00DF91CF911F910F91FF90EF9090
-:10018000DF90CF9008956091840070918500A6E67D
-:10019000B0E0EEE8F0E040E050E00EC02D913D917F
-:1001A000819191912817390729F010928D001092B2
-:1001B0008C0008954F5F5F4F461757077CF30895F3
-:1001C0001F920F920FB60F9211248F939F938091DD
-:1001D000860090918700892B49F080918200909150
-:1001E0008300019690938300809382009F918F916A
-:1001F0000F900FBE0F901F9018951F920F920FB681
-:100200000F9211242F933F938F939F93EF93FF931C
-:1002100021E030E0309389002093880080918C00A9
-:1002200090918D00892B81F08091AA009091AB0074
-:10023000FC01EE0FFF1FE257FF4F31832083019631
-:100240009093AB008093AA00FF91EF919F918F91C3
-:100250003F912F910F900FBE0F901F9018951F92F6
-:100260000F920FB60F9211242F933F938F939F936A
-:10027000EF93FF9322E030E03093890020938800D1
-:1002800080918C0090918D00892B81F08091AA0043
-:100290009091AB00FC01EE0FFF1FE257FF4F31833F
-:1002A000208301969093AB008093AA00FF91EF9179
-:1002B0009F918F913F912F910F900FBE0F901F90A4
-:1002C00018951F920F920FB60F9211242F933F9300
-:1002D0004F935F938F939F93EF93FF9343E050E08F
-:1002E000509389004093880020916400309165000C
-:1002F00020953095309365002093640080918C00A8
-:1003000090918D00019799F42130310581F480910D
-:10031000AA009091AB00FC01EE0FFF1FE257FF4FC8
-:100320005183408301969093AB008093AA00FF9184
-:10033000EF919F918F915F914F913F912F910F90EE
-:100340000FBE0F901F9018951F920F920FB60F922D
-:1003500011242F933F934F935F938F939F93EF932A
-:10036000FF9344E050E0509389004093880020912F
-:10037000640030916500209530953093650020939E
-:10038000640080918C0090918D00019799F4213048
-:10039000310581F48091AA009091AB00FC01EE0F31
-:1003A000FF1FE257FF4F5183408301969093AB00AC
-:1003B0008093AA00FF91EF919F918F915F914F9150
-:1003C0003F912F910F900FBE0F901F901895CF93D4
-:1003D000DF937FD164E070E00BD10196C0E0D0E004
-:1003E000FE01E257FF4F11821082FE01EA59FF4FD2
-:1003F000918380836ED164E070E0FAD001962296FA
-:10040000CC31D10569F7DF91CF9108958FE087BB9B
-:1004100088EE8BBF80E880BD84E084B98AE085BF28
-:1004200083E083BF12BE10928B0010928A0010925C
-:100430008D0010928C00109289001092880010920A
-:10044000AB001092AA001092AD001092AC00109276
-:10045000830010928200109287001092860081E043
-:1004600090E0909365008093640090938500809362
-:10047000840080ED97E09093AF008093AE00789475
-:1004800080E090E02AD1A3DF0895AF92BF92CF928F
-:10049000DF92EF92FF920F931F93CF93DF93B6DF1C
-:1004A00064E0A62EB12C02E010E0C1E0D0E050EDF7
-:1004B000C52E57E0D52E43E0E42EF12C40918A0062
-:1004C00050918B004230510531F1433051053CF4DD
-:1004D0004115510559F041305105A1F716C043307F
-:1004E000510559F14430510569F762C08091880087
-:1004F00090918900892B11F3D0938B00C0938A00CF
-:100500001092890010928800D9CFC6DD10938B001D
-:1005100000938A00D3CF1092AB001092AA00D09320
-:100520006500C0936400D0938D00C0938C00DADD29
-:10053000F0928B00E0928A00C1CF2091AA00309106
-:10054000AB0080918400909185002817390711F540
-:100550001ADE80918C0090918D00009721F0019718
-:1005600009F0ACCF05C0B0928B00A0928A00A6CF54
-:1005700010938B0000938A001092AB001092AA0097
-:1005800080918400909185000196909385008093DE
-:10059000840094CF50938B0040938A008091AC00EC
-:1005A0009091AD0001969093AD008093AC0086CF02
-:1005B00073DD10928900109288001092AB001092A7
-:1005C000AA00D0938500C0938400D0936500C093A7
-:1005D0006400D092AF00C092AE0010928B001092D7
-:1005E0008A00F5DE8091AC009091AD0076D066CFA8
-:1005F00097FB092E07260AD077FD04D00CD006D031
-:1006000000201AF4709561957F4F0895F6F7909544
-:1006100081959F4F0895AA1BBB1B51E107C0AA1FDC
-:10062000BB1FA617B70710F0A61BB70B881F991F93
-:100630005A95A9F780959095BC01CD010895A0E049
-:10064000B0E0E4E2F3E090C0EC01A880B980CA8099
-:10065000DB80A114B104C104D10441F484E2A82ECA
-:1006600089EDB82E8BE5C82E87E0D82EC601B501DE
-:100670002DE133EF41E050E054D027EA31E440E08F
-:1006800050E036D07B018C01C601B5012DE133EF7E
-:1006900041E050E046D0CA01B9012CEE34EF4FEFF3
-:1006A0005FEF26D06E0D7F1D801F911F97FF04C046
-:1006B0006150704080409048688379838A839B832F
-:1006C0009B013F77C901CDB7DEB7EAE069C0B7DF6C
-:1006D000089580E690E0B3DF0895A0E0B0E0809355
-:1006E000600090936100A0936200B093630008954E
-:1006F000FF27EE27BB27AA2760FF04C0A20FB31F66
-:10070000E41FF51F220F331F441F551F9695879531
-:100710007795679589F70097760771F7CF01BD0147
-:10072000089597FB092E05260ED057FD04D04BD017
-:100730000AD0001C38F450954095309521953F4FD4
-:100740004F4F5F4F0895F6F790958095709561959E
-:100750007F4F8F4F9F4F08952F923F924F925F92FE
-:100760006F927F928F929F92AF92BF92CF92DF92C1
-:10077000EF92FF920F931F93CF93DF93CDB7DEB726
-:10078000CA1BDB0B0FB6F894DEBF0FBECDBF0994BA
-:100790002A88398848885F846E847D848C849B8411
-:1007A000AA84B984C884DF80EE80FD800C811B811F
-:1007B000AA81B981CE0FD11D0FB6F894DEBF0FBE4E
-:1007C000CDBFED010895A1E21A2EAA1BBB1BFD01AE
-:1007D0000DC0AA1FBB1FEE1FFF1FA217B307E40720
-:1007E000F50720F0A21BB30BE40BF50B661F771F78
-:1007F000881F991F1A9469F7609570958095909558
-:0E0800009B01AC01BD01CF010895F894FFCF1C
-:04080E0001000000E5
-:00000001FF
